Add `gradient_descent_mse_gp`, a finite time `gp_inference`.
Larger refactor pulling out the predict code to a general utility and simplifying the math in it as well as in the gp_inference function.  Create a new `gaussian_embedding` function to compute the result of mse training at finite time.

Added `compute_var` argument.
Made return argument a namedtuple.

PiperOrigin-RevId: 270297105