Adds the (global) scaled dot-product self-attention layer with option to use multiple attention heads. There are two kinds of attention implemented:
1) Parametric: this is the standard scaled dot-product attention, i.e.,
the dot product between keys and queries is scaled by the squared root
of their dimension. The expression for `nngp`/`ntk` involves an integral
with no known closed form and thus call to `ker_fun` results in an error

2) Fixed: same as Parametric except for scaling the dot products
between keys and queries by their dimension instead of the square root
of the same quantity, and tying the key and query weight matrices.
This makes the `nngp`/`ntk` analytically tractable but for the price
that, unlike in the stochastic case, the dot products of keys
and queries converge to a constant. Because this constant would be zero
if the key and query weights are independent, the variant where these
two weight matrices are tied was implemented resulting in non-constant
attention weights.

PiperOrigin-RevId: 270811553