Support "standard" parameterization for infinite width affine layers for both Dense and Conv layers.
The standard parameterization is an alternative parameterization, which reduces to the standard finite width parameterization at finite width (unlike NTK), and which continues to behave more similarly to the standard finite width parameterization as layer width is taken to infinity, in that relative learning rates between layers and between weights and biases match those of the finite width network.

The core idea is that the width of all layers in a finite width neural network are taken to infinity by multiplying layer width for all layers by a factor s, and dividing pre-activations by a factor 1/sqrt(s). However, unlike in the NTK parameterization, the finite width initialization scale is preserved for all layers (e.g. weights are initialized with variance sigma^2_b / [width]). As a result, the scaling and relative learning rates that stem from the finite width weight and bias initialization scale are preserved.

A short note defining the standard parameterization will be released shortly, and docs will be updated to include a link to that note.

PiperOrigin-RevId: 281828730