Fix `kernel_fn` values and gradients of discontinuous/non-differentiable nonlinearities at 0 - https://github.com/google/neural-tangents/issues/123.
Currently we have NaNs and/or values at 0 that are inconsistent with the limit of finite-width empirical kernels.

First, note that `np.sign(0) == 0`, therefore we must have `T_sign(0, x2) = T_sign(x1, 0) = T_sign(0, 0) = 0`, but we currently have it equal to 1, i.e. we assume incorrectly in our infinite width limit that `np.sign(0) == 1`.

Secondly, JAX defines gradient of ABRelu at 0 to be (a + b) / 2, i.e. mean subgradient. This means that we must have `Tdot_abrelu(0, x2) = Tdot_abrelu(x1, 0) = Tdot_abrelu(0, 0) = [(a + b) / 2]^2`, but we currently have it equal to `(a^2 + b^2) / 2`, which is equivalent to assuming that the gradient is `[(a^2 + b^2) / 2)]^0.5`, i.e. for Relu the gradient at 0 is 1/2^0.5 instead of 1/2.

We fix the above issues by extending `np.arctan2(0, 0) := np.pi / 2` (mathematically the function is undefined, and by default JAX/numpy have it be 0, but `np.pi / 2` gives us correct values above).

Finally, we also extend the gradient of np.arctan2 at (0, 0) to (0, 0). The gradient at 0 is by default undefined, and earlier we had NaN gradients at zero inputs to nonlinearities. While the gradient can't be extended continuously at (0, 0), setting it to (0, 0) at least makes it continuous along `x = 0` or `y = 0`, and helps fix a lot of NaNs.

Also add more tests, including comparisons of gradients of infinite-width kernels with MC estimates.

Make matrix comparison tests fail on NaNs or infinities.

PiperOrigin-RevId: 427073332