Expose the scaling parameter `s` in `parameterization="standard"` (https://arxiv.org/abs/2001.07301). We have had some confusion in the past due to this parameter lacking (https://github.com/google/neural-tangents/issues/132), and this will also allow us to test finite and infinite agreement in the small-N, large-s regime.
Note that in the current solution for each layer the user has to provide a tuple, `(s_{l-1}, s_{l})` in each layer, and ensure themselves that `s_0 == 1` (input to the network is not scaled with `s`), and that `s_{l-1}` passed to the constructor of layer `l` matches `s_{l}` passed to layer `l-1`. If you have better suggestions, please lmk!

PiperOrigin-RevId: 441314378